% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_aggregate_approximation.r
\name{piecewise_aggregate_approximation}
\alias{piecewise_aggregate_approximation}
\title{Piecewise Aggregate Approximation}
\usage{
piecewise_aggregate_approximation(time_series, n_segments = 3,
  na.rm = TRUE)
}
\arguments{
\item{time_series}{Time series}

\item{n_segments}{number of segments}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
The reduced serie
}
\description{
An algorithm for dimensionality reduction,
which the time series is divided into n equal parts and of these parts is taken on average
}
\examples{
 x <- seq(0, pi*2, pi/6)
 ts <- sin(x)*4 + rnorm(length(x))
 ts_norm <- z_normalization(ts)
 plot(x, ts, type="l", col="blue")
 lines(x, ts_norm, type="l", col="red")

 print(paste("Mean is :",mean(ts_norm)))

 print(paste("Sd is:",sd(ts_norm)))
}
\references{
E. Keogh & M. Pazzani. Scaling up dynamic time warping for datamining applications. SIGKDD 2000, pp. 285-289.
}
\author{
Bruno Duru
}
