% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_normalization.r
\name{z_normalization}
\alias{z_normalization}
\title{Z-normalization}
\usage{
z_normalization(timeSeries, na.rm = TRUE)
}
\arguments{
\item{timeSeries}{Time series}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
The time series transformed.
}
\description{
Transform a time series into a time
series with an average of approximately 0 and standard deviation 1.

The formula for the transformation is:
\deqn{x' = (x-\mu)/\sigma}
}
\examples{
 x <- seq(0, pi*2, pi/6)
 ts <- sin(x)*4 + rnorm(length(x))
 ts_norm <- z_normalization(ts)
 plot(x, ts, type="l", col="blue")
 lines(x, ts_norm, type="l", col="red")

 print(paste("Mean is :",mean(ts_norm)))

 print(paste("Sd is:",sd(ts_norm)))

}
\author{
Bruno Duru
}
